model+api/requirements.txt
# $CHA_BEGIN
scikit-learn
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
tensorflow==2.18.0
python-multipart

joblib==1.5.1
Pillow==11.2.1

fastapi
uvicorn

gcloud
google-cloud-storage==2.14.0
google-api-core==2.8.2
googleapis-common-protos==1.56.4

python-dotenv
# $CHA_END
